upload images

+++ src/modules/v1/user/infrastructure/presentation/controller.ts
@@ -1,5 +1,8 @@
-import { Request, Response } from "express";
+import { NextFunction, Request, Response } from "express";
 
+import { Pagination } from "../../../../core/interface/pagination";
+import { Parameters } from "../../../../core/parameters";
+import { validateData } from "../../../../core/presentation/get-errors";
 import { CryptService } from "../../application/services/crypt.service";
 import { UserCreate } from "../../application/user-create";
 import { UserDelete } from "../../application/user-delete";
@@ -7,7 +10,14 @@ import { UserGetByPage } from "../../application/user-get-by-page";
 import { UserGetOne } from "../../application/user-get-one";
 import { UserList } from "../../application/user-list";
 import { UserUpdate } from "../../application/user-update";
+import { User, UserProperties } from "../../domain/roots/user";
 import { UserFactory } from "../../domain/roots/user.factory";
+import {
+  Address,
+  Role,
+  UserCreateValidator,
+} from "./dtos/user-create.validator";
+import { UserGetOneValidator } from "./dtos/user-get-one";
 import { UserDto } from "./dtos/user.response";
 
 export class UserController {
@@ -18,20 +28,62 @@ export class UserController {
     private readonly userUpdate: UserUpdate,
     private readonly userDelete: UserDelete,
     private readonly userGetByPage: UserGetByPage
-  ) {}
-
-  async list(req: Request, res: Response) {
-    const users = await this.userList.execute();
-    res.json(UserDto.fromDomainToResponse(users));
+  ) {
+    this.insert = this.insert.bind(this);
   }
 
-  async insert(req: Request, res: Response) {
-    const body = req.body;
+  async list(req: Request, res: Response) {
+    const result = await this.userList.execute();
+    if (result.isErr()) {
+      return res.status(result.error.status).json({
+        message: result.error.message,
+        stack: result.error.stack,
+      });
+    }
+
+    const users = result.value;
+
+    res.json(UserDto.fromDomainToResponse(users as User[]));
+  }
+
+  async insert(req: Request, res: Response, next: NextFunction) {
+    const body: UserProperties = req.body;
+    const validator = new UserCreateValidator();
+    validator.id = body.id;
+    validator.name = body.name;
+    validator.lastname = body.lastname;
+    validator.email = body.email;
+    validator.password = body.password;
+    validator.roles = body.roles.map((role) => Object.assign(new Role(), role));
+    validator.age = body.age;
+    validator.gender = body.gender;
+    validator.address = Object.assign(new Address(), body.address);
+    validator.image = body.image;
+
+    const isError = await validateData(
+      validator,
+      {
+        whitelist: true,
+      },
+      res
+    );
+    if (!isError) return;
+
     body.password = await CryptService.hash(body.password);
+    if (body.image) body.image = `${Parameters.path_images}/${body.image}`;
 
     const user = UserFactory.create(body);
 
-    const userInserted = await this.userCreate.execute(user);
+    const result = await this.userCreate.execute(user);
+    if (result.isErr()) {
+      return next(result.error);
+      /*return res.status(result.error.status).json({
+        message: result.error.message,
+        stack: result.error.stack,
+      });*/
+    }
+
+    const userInserted = result.value as User;
 
     res.json(UserDto.fromDomainToResponse(userInserted));
   }
@@ -39,7 +91,27 @@ export class UserController {
   async getOne(req: Request, res: Response) {
     const id = req.params.id;
 
-    const user = await this.userGetOne.execute(id);
+    const validator = new UserGetOneValidator();
+    validator.id = id;
+
+    const isError = await validateData(
+      validator,
+      {
+        whitelist: true,
+      },
+      res
+    );
+    if (!isError) return;
+
+    const result = await this.userGetOne.execute(id);
+    if (result.isErr()) {
+      return res.status(result.error.status).json({
+        message: result.error.message,
+        stack: result.error.stack,
+      });
+    }
+
+    const user = result.value as User;
 
     res.json(UserDto.fromDomainToResponse(user));
   }
@@ -48,10 +120,25 @@ export class UserController {
     const id = req.params.id;
     const body = req.body;
 
-    const user = await this.userGetOne.execute(id);
+    const result = await this.userGetOne.execute(id);
+    if (result.isErr()) {
+      return res.status(result.error.status).json({
+        message: result.error.message,
+        stack: result.error.stack,
+      });
+    }
+    const user = result.value as User;
     user.update(body);
 
-    const userUpdated = await this.userUpdate.execute(user);
+    const resultUpdated = await this.userUpdate.execute(user);
+    if (resultUpdated.isErr()) {
+      return res.status(resultUpdated.error.status).json({
+        message: resultUpdated.error.message,
+        stack: resultUpdated.error.stack,
+      });
+    }
+
+    const userUpdated = resultUpdated.value as User;
 
     res.json(UserDto.fromDomainToResponse(userUpdated));
   }
@@ -59,10 +146,26 @@ export class UserController {
   async delete(req: Request, res: Response) {
     const id = req.params.id;
 
-    const user = await this.userGetOne.execute(id);
+    const result = await this.userGetOne.execute(id);
+    if (result.isErr()) {
+      return res.status(result.error.status).json({
+        message: result.error.message,
+        stack: result.error.stack,
+      });
+    }
+
+    const user = result.value as User;
     user.delete();
 
-    const userDeleted = await this.userDelete.execute(user);
+    const resultDelete = await this.userDelete.execute(user);
+    if (resultDelete.isErr()) {
+      return res.status(resultDelete.error.status).json({
+        message: resultDelete.error.message,
+        stack: resultDelete.error.stack,
+      });
+    }
+
+    const userDeleted = resultDelete.value as User;
 
     res.json(UserDto.fromDomainToResponse(userDeleted));
   }
@@ -71,7 +174,15 @@ export class UserController {
     const page = parseInt(req.query.page as string);
     const limit = parseInt(req.query.limit as string);
 
-    const pagination = await this.userGetByPage.execute(page, limit);
+    const result = await this.userGetByPage.execute(page, limit);
+    if (result.isErr()) {
+      return res.status(result.error.status).json({
+        message: result.error.message,
+        stack: result.error.stack,
+      });
+    }
+
+    const pagination = result.value as Pagination<User>;
 
     res.json(UserDto.fromDomainToResponsePagination(pagination));
   }
